Add jbcrypt dependency to pom.xml
A new dependency for jbcrypt has been added to the properties section and the main dependency list in the 'pom.xml' file. With this change, the version of jbcrypt used across the application is controlled in a single place. Also, a hard-coded version number of jbcrypt in 'pinot-common/pom.xml' has been removed to avoid versioning conflicts.